+++ bucketman/app.py
@@ -14,14 +14,30 @@ from bucketman.widgets import (Footer, LocalTree, Prompt, S3BucketSelect,
 
 
 class BucketManApp(textual.app.App):
-    def __init__(self, *args, bucket: str=None, **kwargs):
+    def __init__(self, *args,
+            bucket: str=None,
+            endpoint_url: str=None,
+            access_key_id: str=None,
+            secret_access_key: str=None,
+            **kwargs):
+
+        self.bucket_name = bucket
+
+        session = boto3.Session(
+            aws_access_key_id=access_key_id,
+            aws_secret_access_key=secret_access_key,
+        )
+
+        self.s3_client = session.client('s3', endpoint_url=endpoint_url)
+        self.s3_resource = session.resource('s3', endpoint_url=endpoint_url)
+
         self.left_pane = None
         self.right_pane = None
         self.focused_pane = None
-        self.status_log = None
-        self.directory = None
+        self.status_log = StatusLog()
         self.footer = Footer()
-        self.bucket_name = bucket
+        self.header = Header(style=f"black on {AWS_HEX_COLOR_CODE}")
+
         self.default_bindings = [
             textual.binding.Binding("ctrl+c", "quit", "", allow_forward=False),
             textual.binding.Binding("escape", "quit", "Quit", show=True),
@@ -30,6 +46,18 @@ class BucketManApp(textual.app.App):
         super().__init__(*args, **kwargs)
 
     show_dialog = textual.reactive.Reactive(False)
+
+    @property
+    def left_widget(self):
+        return self.left_pane.window.widget
+
+    @property
+    def right_widget(self):
+        return self.right_pane.window.widget
+
+    @property
+    def focused_widget(self):
+        return self.focused_pane.window.widget
 
     async def load_bindings(self):
         new_bindings = list(self.default_bindings)
@@ -49,34 +77,34 @@ class BucketManApp(textual.app.App):
         await self.load_bindings()
  
     async def action_cycle(self) -> None:
-        if self.left_pane.window.widget == self.focused:
-            await self.right_pane.window.widget.focus()
+        if self.left_widget == self.focused:
+            await self.right_widget.focus()
             await self.load_bindings()
             self.focused_pane = self.right_pane
-        elif self.right_pane.window.widget == self.focused:
-            await self.left_pane.window.widget.focus()
+        elif self.right_widget == self.focused:
+            await self.left_widget.focus()
             await self.load_bindings()
             self.focused_pane = self.left_pane
 
     async def action_delete(self) -> None:
-        if self.left_pane.window.widget == self.focused:
-            path = self.directory.selected_object.path
+        if self.left_widget == self.focused:
+            path = self.left_widget.selected_object.path
             await self.dialog.do_prompt(
                 f"Do you want to delete the path {path}?",
                 self.do_local_delete,
                 path=path
             )
         else:
-            if self.right_pane.window.widget.selected_object.key:
-                prompt = f"Do you want to delete the object {self.right_pane.window.widget.selected_object.key}?"
+            if self.right_widget.selected_object.key:
+                prompt = f"Do you want to delete the object {self.right_widget.selected_object.key}?"
             else:
-                prompt = f"Do you want to empty the bucket {self.right_pane.window.widget.bucket_name}?"
+                prompt = f"Do you want to empty the bucket {self.right_widget.bucket_name}?"
 
             await self.dialog.do_prompt(
                 prompt,
                 self.do_s3_delete,
-                bucket=self.right_pane.window.widget.bucket_name,
-                key=self.right_pane.window.widget.selected_object.key
+                bucket=self.right_widget.bucket_name,
+                key=self.right_widget.selected_object.key
             )
 
     async def do_local_delete(self, path) -> None:
@@ -89,11 +117,10 @@ class BucketManApp(textual.app.App):
             await self.handle_status_update(StatusUpdate(self, message=f'Failed to delete path "{path}": {e.strerror}'))
         else:
             await self.handle_status_update(StatusUpdate(self, message=f'Successfully deleted path "{path}"'))
-            await self.directory.load_objects(self.directory.selected_node.parent)
+            await self.left_widget.load_objects(self.left_widget.selected_node.parent)
 
     async def do_s3_delete(self, bucket, key) -> None:
-        s3 = boto3.resource('s3')
-        bucket = s3.Bucket(bucket)
+        bucket = self.s3_resource.Bucket(bucket)
 
         try:
             bucket.objects.filter(Prefix=key).delete()
@@ -102,25 +129,25 @@ class BucketManApp(textual.app.App):
         except botocore.exceptions.ClientError as e:
             await self.handle_status_update(StatusUpdate(self, message=f'Failed to delete S3 object(s) "{bucket}:{key}": {e.response["Error"]["Message"]}'))
         else:
-            await self.right_pane.window.widget.load_objects(self.right_pane.window.widget.selected_node.parent)
+            await self.right_widget.load_objects(self.right_widget.selected_node.parent)
 
     async def action_copy(self) -> None:
-        if self.left_pane.window.widget == self.focused:
+        if self.left_widget == self.focused:
             message = MakeCopy(
                 self,
                 src_bucket=None,
-                src_path=self.directory.selected_object.path,
-                dst_bucket=self.right_pane.window.widget.bucket_name,
-                dst_path=os.path.dirname(self.right_pane.window.widget.selected_object.key),
+                src_path=self.left_widget.selected_object.path,
+                dst_bucket=self.right_widget.bucket_name,
+                dst_path=os.path.dirname(self.right_widget.selected_object.key),
                 recursive=True
             )
         else:
             message = MakeCopy(
                 self,
-                src_bucket=self.right_pane.window.widget.bucket_name,
-                src_path=self.right_pane.window.widget.selected_object.key,
+                src_bucket=self.right_widget.bucket_name,
+                src_path=self.right_widget.selected_object.key,
                 dst_bucket=None,
-                dst_path=self.directory.data.path,
+                dst_path=self.left_widget.data.path,
                 recursive=True
             )
 
@@ -131,20 +158,20 @@ class BucketManApp(textual.app.App):
         elif message.src_bucket is None and message.dst_bucket is not None:
             key = os.path.join(message.dst_path, os.path.basename(message.src_path))
             await self.dialog.do_prompt(
-                f"Do you want to upload the file {message.src_path} to {self.right_pane.window.widget.bucket_name}/{key}?",
+                f"Do you want to upload the file {message.src_path} to {self.right_widget.bucket_name}/{key}?",
                 self.do_upload,
                 path=message.src_path,
-                bucket=self.right_pane.window.widget.bucket_name,
+                bucket=self.right_widget.bucket_name,
                 key=key
             )
         # download
         elif message.src_bucket is not None and message.dst_bucket is None:
             path = os.path.join(message.dst_path, os.path.basename(message.src_path))
             await self.dialog.do_prompt(
-                f"Do you want to download the object {self.right_pane.window.widget.bucket_name}/{self.right_pane.window.widget.selected_object.key} to {path}?",
+                f"Do you want to download the object {self.right_widget.bucket_name}/{self.right_widget.selected_object.key} to {path}?",
                 self.do_download,
-                bucket=self.right_pane.window.widget.bucket_name,
-                key=self.right_pane.window.widget.selected_object.key,
+                bucket=self.right_widget.bucket_name,
+                key=self.right_widget.selected_object.key,
                 path=path
             )
         # copy from one bucket to another
@@ -152,33 +179,30 @@ class BucketManApp(textual.app.App):
             await self.handle_status_update(StatusUpdate(self, message='Copying from S3 to S3 is not yet supported!'))
 
     async def do_upload(self, path, bucket, key) -> None:
-        client = boto3.client('s3')
-
         try:
             if os.path.isdir(path):
                 for root, dirs, files in os.walk(path):
                     for file in files:
                         src = os.path.join(root, file)
                         dst = os.path.normpath(os.path.join(key, os.path.relpath(root, path),file))
-                        client.upload_file(src, bucket, dst)
+                        self.s3_client.upload_file(src, bucket, dst)
                         await self.handle_status_update(StatusUpdate(self, message=f'Uploaded file {src} to {bucket}/{dst}'))
             else:
-                client.upload_file(path, bucket, key)
+                self.s3_client.upload_file(path, bucket, key)
         except botocore.exceptions.ClientError as e:
             await self.handle_status_update(StatusUpdate(self, message=f'Failed to upload file {path} to {bucket}/{key}: {e.response["Error"]["Message"]}'))
         else:
-            node_to_reload = self.right_pane.window.widget.selected_node.parent if self.right_pane.window.widget.selected_node.parent else self.right_pane.window.widget.selected_node
-            await self.right_pane.window.widget.load_objects(node_to_reload)
+            node_to_reload = self.right_widget.selected_node.parent if self.right_widget.selected_node.parent else self.right_widget.selected_node
+            await self.right_widget.load_objects(node_to_reload)
 
     async def do_download(self, bucket, key, path) -> None:
-        client = boto3.client('s3')
-
         try:
-            client.download_file(bucket, key, path)
+            self.s3_client.download_file(bucket, key, path)
         except botocore.exceptions.ClientError as e:
             await self.handle_status_update(StatusUpdate(self, message=f'Failed to download object {bucket}/{key} to {path}: {e.response["Error"]["Message"]}'))
         else:
-            await self.directory.load_objects(self.directory.selected_node.parent)
+            node_to_reload = self.left_widget.selected_node.parent if self.left_widget.selected_node.parent else self.left_widget.selected_node
+            await self.left_widget.load_objects(node_to_reload)
             await self.handle_status_update(StatusUpdate(self, message=f'Downloaded object {bucket}/{key} to {path}'))
 
     async def handle_status_update(self, message: StatusUpdate) -> None:
@@ -192,20 +216,20 @@ class BucketManApp(textual.app.App):
         if self.show_dialog:
             await self.dialog.focus()
         else:
-            await self.focused_pane.window.widget.focus()
+            await self.focused_widget.focus()
 
     async def show_select_bucket(self) -> None:
         await self.focused_pane.update(
             S3BucketSelect(self.bucket_selected)
         )
-        await self.focused_pane.window.widget.focus()
+        await self.focused_widget.focus()
         await self.load_bindings()
 
     async def bucket_selected(self):
-        bucket_name = self.focused_pane.window.widget.selected_bucket
+        bucket_name = self.focused_widget.selected_bucket
         s3tree = S3Tree(bucket_name, name="s3")
         await self.focused_pane.update(s3tree)
-        await self.focused_pane.window.widget.focus()
+        await self.focused_widget.focus()
         await self.load_bindings()
 
     async def on_mount(self) -> None:
@@ -213,10 +237,9 @@ class BucketManApp(textual.app.App):
         await self.view.dock(self.dialog, edge="bottom", size=20, z=1)
         self.dialog.layout_offset_y = 20
 
-        self.status_log = StatusLog()
-        self.directory = LocalTree(os.getcwd())
-        self.directory.show_cursor = True
-        self.left_pane = textual.widgets.ScrollView(self.directory, name="left_pane")
+        directory = LocalTree(os.getcwd())
+        directory.show_cursor = True
+        self.left_pane = textual.widgets.ScrollView(directory, name="left_pane")
         if self.bucket_name:
             widget = S3Tree(self.bucket_name, name="s3")
         else:
@@ -244,7 +267,7 @@ class BucketManApp(textual.app.App):
             footer="left-start|right-end,footer",
         )
 
-        grid.place(header=Header(style=f"black on {AWS_HEX_COLOR_CODE}"))
+        grid.place(header=self.header)
         grid.place(left=self.left_pane)
         grid.place(middle=VerticalDivider())
         grid.place(right=self.right_pane)